server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
google-genai==0.5.0
torch==2.1.2
transformers==4.36.2
Pillow==10.2.0
httpx==0.25.2
email-validator==2.3.1
